+++ django_goodies/__init__.py
@@ -1,1 +1,4 @@
 __version__ = '0.3'
+
+# Shortcuts to goodies
+from django_goodies.misc import AjaxResponse  # NOQA