+++ urls.py
@@ -1,2 +1,0 @@
-# Unfortunately required by ROOT_URLCONF setting
-urlpatterns = []