+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Change Log
 0.4
 ===
 
+* Added AjaxResponse class
 * Added GOODIES_COMMON_INFO_REQUIRE_USER_ON_SAVE setting
 
 0.3